requirements.txt
streamlit
databricks-sql-connector
databricks-sdk
pandas
plotly
